requirements.txt
Flask
flatlib
pyswisseph
geopy
requests
pytz
openai
gunicorn